# assistants/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Symbia Assistants Backend API",
    "version": "1.0.0",
    "description": "Backend APIs for prompt graphs, actor principals, run orchestration, rules engine, and LLM settings."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Identity Service"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key from Identity Service"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Session cookie from Identity Service"
      }
    },
...
  },
  "paths": {
...
    "/assistants-admin": {
      "get": {
        "tags": [
          "assistants-admin"
        ],
        "summary": "List custom assistants",
        "responses": {
          "200": {
            "description": "List of custom assistants"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "post": {
        "tags": [
          "assistants-admin"
        ],
        "summary": "Create custom assistant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "key",
                  "name"
                ],
                "properties": {
                  "key": {
                    "type": "string",
                    "pattern": "^[a-z0-9-]+$"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "capabilities": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive",
                      "draft"
                    ]
                  },
                  "systemPrompt": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "temperature": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assistant created"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Assistant with key already exists"
          }
        }
      }
    },
    "/assistants-admin/{key}": {
      "get": {
        "tags": [
          "assistants-admin"
        ],
        "summary": "Get custom assistant",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assistant details"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Assistant not found"
          }
        }
      },
      "put": {
        "tags": [
          "assistants-admin"
        ],
        "summary": "Update custom assistant",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assistant updated"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Assistant not found"
          }
        }
      },
      "delete": {
        "tags": [
          "assistants-admin"
        ],
        "summary": "Delete custom assistant",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assistant deleted"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Assistant not found"
          }
        }
      }
    }
  }
}